% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnector.R
\name{DatabaseConnector}
\alias{DatabaseConnector}
\title{R6 Class representing a database connection.}
\description{
R6 Class representing a database connection.

R6 Class representing a database connection.
}
\details{
This class provides methods to initialize a database connection either using
a configuration file or direct parameters, and retrieve views from the database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{A database connection object created using pool::dbPool.
Initialize the database connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseConnector-new}{\code{DatabaseConnector$new()}}
\item \href{#method-DatabaseConnector-get_view}{\code{DatabaseConnector$get_view()}}
\item \href{#method-DatabaseConnector-finalize}{\code{DatabaseConnector$finalize()}}
\item \href{#method-DatabaseConnector-clone}{\code{DatabaseConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$new(
  config_path = NULL,
  dbname = NULL,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL,
  sslmode = "require",
  sslrootcert = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_path}}{Path to the configuration file.}

\item{\code{dbname}}{Database name.}

\item{\code{host}}{Database host.}

\item{\code{port}}{Database port.}

\item{\code{user}}{Database user.}

\item{\code{password}}{Database password.}

\item{\code{sslmode}}{SSL mode for the database connection.}

\item{\code{sslrootcert}}{Path to the SSL root certificate.
Retrieve a view from the database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-get_view"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-get_view}{}}}
\subsection{Method \code{get_view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$get_view(view_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_name}}{Name of the view to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tbl object representing the specified view.
Finalize the database connection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes the database connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
